feat: view academics report (#32)

+++ utils/utilities.h
@@ -37,7 +37,7 @@ void validateSubjectMarks(short marks[], short index)
 }
 
 // Validate number of days
-void validateDays(short &days)
+void validateDays(float &days)
 {
   while (days > 366 || days < 1)
   {
